Allow specifying a transaction size when projecting multiple events
The process_events method may be called with 1 or more events. It then
loops through each event, opens a transaction and processes the event.

This change allows specifying the number of events to process before
closing the transaction.

Processing more events inside the transactions should lead to speed
increases in most cases.